+++ packages/engine/Source/DataSources/PositionPropertyArray.js
@@ -3,6 +3,7 @@ import defined from "../Core/defined.js";
 import DeveloperError from "../Core/DeveloperError.js";
 import Event from "../Core/Event.js";
 import EventHelper from "../Core/EventHelper.js";
+import JulianDate from "../Core/JulianDate.js";
 import ReferenceFrame from "../Core/ReferenceFrame.js";
 import Property from "./Property.js";
 
@@ -76,14 +77,19 @@ Object.defineProperties(PositionPropertyArray.prototype, {
   },
 });
 
+const timeScratch = new JulianDate();
+
 /**
  * Gets the value of the property.
  *
- * @param {JulianDate} time The time for which to retrieve the value.
+ * @param {JulianDate} [time=JulianDate.now()] The time for which to retrieve the value. If omitted, the current system time is used.
  * @param {Cartesian3[]} [result] The object to store the value into, if omitted, a new instance is created and returned.
  * @returns {Cartesian3[]} The modified result parameter or a new instance if the result parameter was not supplied.
  */
 PositionPropertyArray.prototype.getValue = function (time, result) {
+  if (!defined(time)) {
+    time = JulianDate.now(timeScratch);
+  }
   return this.getValueInReferenceFrame(time, ReferenceFrame.FIXED, result);
 };