+++ packages/engine/Source/DataSources/Property.js
@@ -52,7 +52,7 @@ Object.defineProperties(Property.prototype, {
  * Gets the value of the property at the provided time.
  * @function
  *
- * @param {JulianDate} time The time for which to retrieve the value.
+ * @param {JulianDate} [time=JulianDate.now()] The time for which to retrieve the value. If omitted, the current system time is used.
  * @param {object} [result] The object to store the value into, if omitted, a new instance is created and returned.
  * @returns {object} The modified result parameter or a new instance if the result parameter was not supplied.
  */