+++ packages/engine/Source/Renderer/Context.js
@@ -160,7 +160,7 @@ function Context(canvas, options) {
     "OES_texture_half_float_linear",
   ]);
 
-  this._textureLod = !!getExtension(gl, ["EXT_shader_texture_lod"]);
+  this._supportsTextureLod = !!getExtension(gl, ["EXT_shader_texture_lod"]);
 
   this._colorBufferFloat = !!getExtension(gl, [
     "EXT_color_buffer_float",
@@ -784,11 +784,11 @@ Object.defineProperties(Context.prototype, {
    * access to explicit LOD selection in texture sampling functions.
    * @memberof Context.prototype
    * @type {boolean}
-   * @see {@link https://www.khronos.org/registry/webgl/extensions/OES_texture_float_linear/}
+   * @see {@link https://registry.khronos.org/webgl/extensions/EXT_shader_texture_lod/}
    */
-  textureLod: {
+  supportsTextureLod: {
     get: function () {
-      return this._webgl2 || this._textureFloatLinear;
+      return this._webgl2 || this._supportsTextureLod;
     },
   },