+++ packages/engine/Source/Renderer/CubeMap.js
@@ -21,10 +21,31 @@ import TextureMinificationFilter from "./TextureMinificationFilter.js";
 import VertexArray from "./VertexArray.js";
 
 /**
+ * @typedef CubeMap.BufferSource
+ *
+ * @property {TypedArray} arrayBufferView A view of a binary data buffer containing pixel values.
+ * @property {number} width The width of one face of the cube map, in pixels. Must be equal to height.
+ * @property {number} height The height of one face of the cube map, in pixels. Must be equal to width.
+ *
+ * @private
+ */
+
+/**
+ * @typedef CubeMap.Source
+ *
+ * @property {ImageData|HTMLImageElement|HTMLCanvasElement|HTMLVideoElement|CubeMap.BufferSource} positiveX
+ * @property {ImageData|HTMLImageElement|HTMLCanvasElement|HTMLVideoElement|CubeMap.BufferSource} negativeX
+ * @property {ImageData|HTMLImageElement|HTMLCanvasElement|HTMLVideoElement|CubeMap.BufferSource} positiveY
+ * @property {ImageData|HTMLImageElement|HTMLCanvasElement|HTMLVideoElement|CubeMap.BufferSource} negativeY
+ * @property {ImageData|HTMLImageElement|HTMLCanvasElement|HTMLVideoElement|CubeMap.BufferSource} positiveZ
+ * @property {ImageData|HTMLImageElement|HTMLCanvasElement|HTMLVideoElement|CubeMap.BufferSource} negativeZ
+ */
+
+/**
  * @typedef CubeMap.ConstructorOptions
  *
  * @property {Context} context
- * @property {object} [source] The source for texel values to be loaded into the texture
+ * @property {CubeMap.Source} [source] The source for texel values to be loaded into the texture.
  * @property {PixelFormat} [pixelFormat=PixelFormat.RGBA] The format of each pixel, i.e., the number of components it has and what they represent.
  * @property {PixelDatatype} [pixelDatatype=PixelDatatype.UNSIGNED_BYTE] The data type of each pixel.
  * @property {boolean} [flipY=true] If true, the source values will be read as if the y-axis is inverted (y=0 at the top).
@@ -73,22 +94,26 @@ function CubeMap(options) {
 
   let { width, height } = options;
 
-  const faceNames = CubeMap.faceNames;
-
   if (defined(source)) {
     //>>includeStart('debug', pragmas.debug);
-    if (!faceNames.every((faceName) => defined(source[faceName]))) {
+    if (
+      !Object.values(CubeMap.FaceName).every((faceName) =>
+        defined(source[faceName])
+      )
+    ) {
       throw new DeveloperError(
-        `options.source requires faces ${faceNames.join(", ")}.`
+        `options.source requires faces ${Object.values(CubeMap.FaceName).join(
+          ", "
+        )}.`
       );
     }
     //>>includeEnd('debug');
 
-    ({ width, height } = source[faceNames[0]]);
+    ({ width, height } = source.positiveX);
 
     //>>includeStart('debug', pragmas.debug);
-    for (let i = 1; i < 6; ++i) {
-      const face = source[faceNames[i]];
+    for (const faceName of CubeMap.faces()) {
+      const face = source[faceName];
       if (Number(face.width) !== width || Number(face.height) !== height) {
         throw new DeveloperError(
           "Each face in options.source must have the same width and height."
@@ -213,22 +238,12 @@ function CubeMap(options) {
     );
   }
 
-  for (let i = 0; i < faceNames.length; i++) {
-    const faceName = faceNames[i];
+  for (const faceName of CubeMap.faces()) {
     loadFace(this[faceName], source?.[faceName], 0);
   }
 
   gl.bindTexture(textureTarget, null);
 }
-
-CubeMap.faceNames = Object.freeze([
-  "positiveX",
-  "negativeX",
-  "positiveY",
-  "negativeY",
-  "positiveZ",
-  "negativeZ",
-]);
 
 /**
  * TODO
@@ -258,76 +273,53 @@ CubeMap.prototype.copyFace = function (frameState, texture, face, mipLevel) {
 };
 
 /**
- * Creates a CubeMap, using a texel data source that includes pre-generated mipmaps.
+ * An enum defining the names of the faces of a cube map.
  *
- * @param {CubeMap.ConstructorOptions} options An object describing initialization options.
- * @returns {CubeMap}
+ * @alias {CubeMap.FaceName}
+ * @enum {string}
  *
  * @private
  */
-CubeMap.fromMipmaps = function (options) {
-  options = defaultValue(options, defaultValue.EMPTY_OBJECT);
-  const { source, skipColorSpaceConversion } = options;
+CubeMap.FaceName = Object.freeze({
+  POSITIVEX: "positiveX",
+  NEGATIVEX: "negativeX",
+  POSITIVEY: "positiveY",
+  NEGATIVEY: "negativeY",
+  POSITIVEZ: "positiveZ",
+  NEGATIVEZ: "negativeZ",
+});
 
-  //>>includeStart('debug', pragmas.debug);
-  Check.defined("options.source", source);
+function* makeFacesIterator() {
+  yield CubeMap.FaceName.POSITIVEX;
+  yield CubeMap.FaceName.NEGATIVEX;
+  yield CubeMap.FaceName.POSITIVEY;
+  yield CubeMap.FaceName.NEGATIVEY;
+  yield CubeMap.FaceName.POSITIVEZ;
+  yield CubeMap.FaceName.NEGATIVEZ;
+}
 
-  if (!Array.isArray(source)) {
-    throw new DeveloperError(`options.source must be an array`);
-  }
-  const faceSize = source[0].positiveX.width;
-  const mipCount = Math.log2(faceSize) + 1;
-  if (source.length !== mipCount) {
-    throw new DeveloperError(`all mip levels must be defined`);
-  }
-  // TODO: Verify that the structure of each mip level matches the first.
-  //>>includeEnd('debug');
-
-  options.source = source[0];
-  const cubeMap = new CubeMap(options);
-
-  const gl = cubeMap._context._gl;
-  const texture = cubeMap._texture;
-  const textureTarget = cubeMap._textureTarget;
-
-  gl.activeTexture(gl.TEXTURE0);
-  gl.bindTexture(textureTarget, texture);
-
-  if (skipColorSpaceConversion) {
-    gl.pixelStorei(gl.UNPACK_COLORSPACE_CONVERSION_WEBGL, gl.NONE);
-  } else {
-    gl.pixelStorei(
-      gl.UNPACK_COLORSPACE_CONVERSION_WEBGL,
-      gl.BROWSER_DEFAULT_WEBGL
-    );
-  }
-
-  const faceNames = CubeMap.faceNames;
-  for (let mipLevel = 1; mipLevel < source.length; mipLevel++) {
-    const mipSource = source[mipLevel];
-    for (let j = 0; j < faceNames.length; j++) {
-      const faceName = faceNames[j];
-      loadFace(cubeMap[faceName], mipSource[faceName], mipLevel);
-    }
-  }
-
-  gl.bindTexture(textureTarget, null);
-
-  cubeMap._hasMipmap = true;
-
-  return cubeMap;
+/**
+ * Creates an iterator for looping over the cubemap faces.
+ *
+ * @type {Iterable<CubeMap.FaceName>}
+ *
+ * @private
+ */
+CubeMap.faces = function () {
+  return makeFacesIterator();
 };
 
 /**
  * Load texel data into one face of a cube map.
  *
  * @param {CubeMapFace} cubeMapFace The face to which texel values will be loaded.
- * @param {object} [source] The source for texel values to be loaded into the texture.
- * @param {number} mipLevel The mip level to which the texel values will be loaded.
+ * @param {ImageData|HTMLImageElement|HTMLCanvasElement|HTMLVideoElement|CubeMap.BufferSource} [source] The source for texel values to be loaded into the texture.
+ * @param {number} [mipLevel=0] The mip level to which the texel values will be loaded.
  *
  * @private
  */
 function loadFace(cubeMapFace, source, mipLevel) {
+  mipLevel = defaultValue(mipLevel, 0);
   const targetFace = cubeMapFace._targetFace;
   const size = Math.max(Math.floor(cubeMapFace._size / 2 ** mipLevel), 1);
   const pixelFormat = cubeMapFace._pixelFormat;
@@ -353,7 +345,7 @@ function loadFace(cubeMapFace, source, mipLevel) {
     return;
   }
 
-  let { arrayBufferView = source.bufferView } = source;
+  let { arrayBufferView } = source;
 
   let unpackAlignment = 4;
   if (defined(arrayBufferView)) {
@@ -491,59 +483,25 @@ Object.defineProperties(CubeMap.prototype, {
   },
 });
 
-function* makeFacesIterator(cubeMap) {
-  yield cubeMap._negativeX;
-  yield cubeMap._negativeY;
-  yield cubeMap._negativeZ;
-  yield cubeMap._positiveX;
-  yield cubeMap._positiveY;
-  yield cubeMap._positiveZ;
-}
-
-/**
- * Creates an iterator for looping over the cubemap faces.
- * @type {Iterable<CubeMapFace>}
- */
-CubeMap.prototype.faces = function () {
-  return makeFacesIterator(this);
-};
-
 /**
  * TODO
- * @param {*} face
- * @param {*} result
+ * @param {CubeMap.FaceName} face
+ * @param {Cartesian3} result
  * @returns
  */
-CubeMap.prototype.getDirection = function (face, result) {
-  switch (face) {
-    case this.positiveX:
-      return Cartesian3.clone(Cartesian3.UNIT_X, result);
-    case this.negativeX:
-      return Cartesian3.negate(Cartesian3.UNIT_X, result);
-    case this.positiveY:
-      return Cartesian3.clone(Cartesian3.UNIT_Y, result);
-    case this.negativeY:
-      return Cartesian3.negate(Cartesian3.UNIT_Y, result);
-    case this.positiveZ:
-      return Cartesian3.clone(Cartesian3.UNIT_Z, result);
-    case this.negativeZ:
-      return Cartesian3.negate(Cartesian3.UNIT_Z, result);
-  }
-};
-
 CubeMap.getDirection = function (face, result) {
   switch (face) {
-    case CubeMap.faceNames[0]:
+    case CubeMap.FaceName.POSITIVEX:
       return Cartesian3.clone(Cartesian3.UNIT_X, result);
-    case CubeMap.faceNames[1]:
+    case CubeMap.FaceName.NEGATIVEX:
       return Cartesian3.negate(Cartesian3.UNIT_X, result);
-    case CubeMap.faceNames[2]:
+    case CubeMap.FaceName.POSITIVEY:
       return Cartesian3.clone(Cartesian3.UNIT_Y, result);
-    case CubeMap.faceNames[3]:
+    case CubeMap.FaceName.NEGATIVEY:
       return Cartesian3.negate(Cartesian3.UNIT_Y, result);
-    case CubeMap.faceNames[4]:
+    case CubeMap.FaceName.POSITIVEZ:
       return Cartesian3.clone(Cartesian3.UNIT_Z, result);
-    case CubeMap.faceNames[5]:
+    case CubeMap.FaceName.NEGATIVEZ:
       return Cartesian3.negate(Cartesian3.UNIT_Z, result);
   }
 };
@@ -598,6 +556,57 @@ function setupSampler(cubeMap, sampler) {
   }
   gl.bindTexture(target, null);
 }
+
+/**
+ * Load a complete mipmap chain for each cubemap face.
+ *
+ * @param {CubeMap.Source[]} source The source data for each mip level, beginning at level 1.
+ * @param {boolean} [skipColorSpaceConversion=false] If true, color space conversions will be skipped when reading the texel values.
+ *
+ * @private
+ */
+CubeMap.prototype.loadMipmaps = function (source, skipColorSpaceConversion) {
+  //>>includeStart('debug', pragmas.debug);
+  Check.defined("source", source);
+  if (!Array.isArray(source)) {
+    throw new DeveloperError(`source must be an array`);
+  }
+  const mipCount = Math.log2(this._size);
+  if (source.length !== mipCount) {
+    throw new DeveloperError(`all mip levels must be defined`);
+  }
+  //>>includeEnd('debug');
+
+  skipColorSpaceConversion = defaultValue(skipColorSpaceConversion, false);
+  const gl = this._context._gl;
+  const texture = this._texture;
+  const textureTarget = this._textureTarget;
+
+  gl.activeTexture(gl.TEXTURE0);
+  gl.bindTexture(textureTarget, texture);
+
+  if (skipColorSpaceConversion) {
+    gl.pixelStorei(gl.UNPACK_COLORSPACE_CONVERSION_WEBGL, gl.NONE);
+  } else {
+    gl.pixelStorei(
+      gl.UNPACK_COLORSPACE_CONVERSION_WEBGL,
+      gl.BROWSER_DEFAULT_WEBGL
+    );
+  }
+
+  for (let i = 0; i < source.length; i++) {
+    const mipSource = source[i];
+    // mipLevel 0 was the base layer, already loaded when the CubeMap was constructed.
+    const mipLevel = i + 1;
+    for (const faceName of CubeMap.faces()) {
+      loadFace(this[faceName], mipSource[faceName], mipLevel);
+    }
+  }
+
+  gl.bindTexture(textureTarget, null);
+
+  this._hasMipmap = true;
+};
 
 /**
  * Generates a complete mipmap chain for each cubemap face.