+++ packages/engine/Source/Renderer/loadCubeMap.js
@@ -42,7 +42,9 @@ function loadCubeMap(context, urls, skipColorSpaceConversion) {
   //>>includeStart('debug', pragmas.debug);
   Check.defined("context", context);
   Check.defined("urls", urls);
-  if (CubeMap.faceNames.some((faceName) => !defined(urls[faceName]))) {
+  if (
+    Object.values(CubeMap.FaceName).some((faceName) => !defined(urls[faceName]))
+  ) {
     throw new DeveloperError(
       "urls must have positiveX, negativeX, positiveY, negativeY, positiveZ, and negativeZ properties."
     );