+++ CHANGES.md
@@ -2,8 +2,20 @@
 
 ### 1.121 - 2024-09-01
 
+#### @cesium/engine
+
+##### Additions :tada:
+
+- Made the `time` parameter optional for `Property`, using `JulianDate.now()` as default. [#12099](https://github.com/CesiumGS/cesium/pull/12099)
+
+##### Fixes :wrench:
+
+- Fixed cube-mapping artifacts in image-based lighting. [#12100](https://github.com/CesiumGS/cesium/pull/12100)
+- Fixed specular reflection artifact in PBR direct lighting. [#12116](https://github.com/CesiumGS/cesium/pull/12116)
+
 ##### Breaking Changes :mega:
 
+- Custom specular environment maps in `ImageBasedLighting` now require either a WebGL2 context or a WebGL1 context that supports the [`EXT_shader_texture_lod` extension](https://registry.khronos.org/webgl/extensions/EXT_shader_texture_lod/).
 - `ImageBasedLighting.luminanceAtZenith` has been removed. Use newFoo instead.
 
 ### 1.120 - 2024-08-01